inc_dec_step_one.c: use long long for the 64-bit step
`long` is 32-bit on LLP64 (Windows), where `9999999999L` does not fit, so
the fixture failed the PE/aarch64 parity lane while passing on LP64. Use
`long long`, which is 64-bit on both models.

+++ tests/fixtures/c/inc_dec_step_one.c
@@ -6,8 +6,10 @@
 
 static int plus_one(int x) { return x + 1; }
 static int minus_one(int x) { return x - 1; }
-static long plus_one_l(long x) { return x + 1; }
-static long minus_neg_one(long x) { return x - (-1); }
+// `long long` is 64-bit on LP64 and LLP64 alike; plain `long` is 32-bit
+// on Windows, where the wide literal below would not fit.
+static long long plus_one_l(long long x) { return x + 1; }
+static long long minus_neg_one(long long x) { return x - (-1); }
 
 static int count_up(int n) {
     int c = 0;
@@ -24,7 +26,7 @@ static unsigned wrap(unsigned x) {
 int main(void) {
     if (plus_one(41) != 42) return 1;
     if (minus_one(43) != 42) return 2;
-    if (plus_one_l(9999999999L) != 10000000000L) return 3;
+    if (plus_one_l(9999999999LL) != 10000000000LL) return 3;
     if (minus_neg_one(41) != 42) return 4;
     if (count_up(42) != 42) return 5;
     if (wrap(0xffffffffu) != 0u) return 6;
